created global types/interfaces file
