+++ client/src/types.ts
@@ -1,0 +1,12 @@
+export interface IEvent {
+    name: string;
+    start: Date;
+    end: Date;
+}
+
+export interface ITask {
+    name: string;
+    description: string;
+    tags: string[];
+    deadline?: Date;
+}
